perf: load the code editor on demand
136KB off every dashboard page, including the light ones.

CodeMirror plus its language modes is the single heaviest thing in the bundle -
464KB across two chunks, more than React and the Next runtime combined. Thirty-
six components import CodeEditor, and every one of them renders it inside a tab,
a dialog or a settings panel that most visits never open.

Rather than touching all thirty-six, the implementation moved to
code-editor-impl.tsx and the public name became a `next/dynamic` wrapper. Call
sites are unchanged.

    application  2614K -> 2478K
    compose      2541K -> 2405K
    postgres     2262K -> 2126K
    redis        2060K -> 1924K

Cumulative for the application page across this branch: 3171K -> 2478K, a 22%
reduction.

The saving is 136KB rather than the full 464KB because part of CodeMirror is
still reachable from chunks the pages load for other reasons. Measured, not
assumed - which is why the number here is the smaller one.
